+++ libicp/src/icpPointToPoint.cpp
@@ -56,15 +56,15 @@ double IcpPointToPoint::fitStep (double *T,const int32_t T_num,Matrix &R,Matrix 
 #pragma omp parallel for private(i) default(none) shared(T,active,nact,p_m,p_t,r00,r01,r10,r11,t0,t1) reduction(+:mum0,mum1, mut0,mut1) // schedule (dynamic,2)
     for (i=0; i<nact; i++) {
       // kd tree query + result
-      std::vector<float>         query(m_dim);
+      std::vector<double>         query(m_dim);
       kdtree::KDTreeResultVector result;
   
       // get index of active point
       int32_t idx = active[i];
 
       // transform point according to R|t
-      query[0] = (float)(r00*T[idx*2+0] + r01*T[idx*2+1] + t0);
-      query[1] = (float)(r10*T[idx*2+0] + r11*T[idx*2+1] + t1);
+      query[0] = (r00*T[idx*2+0] + r01*T[idx*2+1] + t0);
+      query[1] = (r10*T[idx*2+0] + r11*T[idx*2+1] + t1);
 
       // search nearest neighbor
       m_kd_tree->n_nearest(query,1,result);
@@ -98,16 +98,16 @@ double IcpPointToPoint::fitStep (double *T,const int32_t T_num,Matrix &R,Matrix 
 #pragma omp parallel for private(i) default(none) shared(T,active,nact,p_m,p_t,r00,r01,r02,r10,r11,r12,r20,r21,r22,t0,t1,t2) reduction(+:mum0,mum1,mum2, mut0,mut1,mut2) // schedule (dynamic,2)
     for (i=0; i<nact; i++) {
       // kd tree query + result
-      std::vector<float>         query(m_dim);
+      std::vector<double>         query(m_dim);
       kdtree::KDTreeResultVector result;
 
       // get index of active point
       int32_t idx = active[i];
 
       // transform point according to R|t
-      query[0] = (float)(r00*T[idx*3+0] + r01*T[idx*3+1] + r02*T[idx*3+2] + t0);
-      query[1] = (float)(r10*T[idx*3+0] + r11*T[idx*3+1] + r12*T[idx*3+2] + t1);
-      query[2] = (float)(r20*T[idx*3+0] + r21*T[idx*3+1] + r22*T[idx*3+2] + t2);
+      query[0] = (r00*T[idx*3+0] + r01*T[idx*3+1] + r02*T[idx*3+2] + t0);
+      query[1] = (r10*T[idx*3+0] + r11*T[idx*3+1] + r12*T[idx*3+2] + t1);
+      query[2] = (r20*T[idx*3+0] + r21*T[idx*3+1] + r22*T[idx*3+2] + t2);
 
       // search nearest neighbor
       m_kd_tree->n_nearest(query,1,result);
@@ -165,7 +165,7 @@ std::vector<int32_t> IcpPointToPoint::getInliers (double *T,const int32_t T_num,
 
   // init inlier vector + query point + query result
   vector<int32_t>            inliers;
-  std::vector<float>         query(m_dim);
+  std::vector<double>         query(m_dim);
   kdtree::KDTreeResultVector neighbor;
   
   // dimensionality 2
@@ -180,8 +180,8 @@ std::vector<int32_t> IcpPointToPoint::getInliers (double *T,const int32_t T_num,
     for (int32_t i=0; i<T_num; i++) {
 
       // transform point according to R|t
-      query[0] = (float)(r00*T[i*2+0] + r01*T[i*2+1] + t0);
-      query[1] = (float)(r10*T[i*2+0] + r11*T[i*2+1] + t1);
+      query[0] = (r00*T[i*2+0] + r01*T[i*2+1] + t0);
+      query[1] = (r10*T[i*2+0] + r11*T[i*2+1] + t1);
 
       // search nearest neighbor
       m_kd_tree->n_nearest(query,1,neighbor);
@@ -204,9 +204,9 @@ std::vector<int32_t> IcpPointToPoint::getInliers (double *T,const int32_t T_num,
     for (int32_t i=0; i<T_num; i++) {
 
       // transform point according to R|t
-      query[0] = (float)(r00*T[i*3+0] + r01*T[i*3+1] + r02*T[i*3+2] + t0);
-      query[1] = (float)(r10*T[i*3+0] + r11*T[i*3+1] + r12*T[i*3+2] + t1);
-      query[2] = (float)(r20*T[i*3+0] + r21*T[i*3+1] + r22*T[i*3+2] + t2);
+      query[0] = (r00*T[i*3+0] + r01*T[i*3+1] + r02*T[i*3+2] + t0);
+      query[1] = (r10*T[i*3+0] + r11*T[i*3+1] + r12*T[i*3+2] + t1);
+      query[2] = (r20*T[i*3+0] + r21*T[i*3+1] + r22*T[i*3+2] + t2);
 
       // search nearest neighbor
       m_kd_tree->n_nearest(query,1,neighbor);
@@ -234,15 +234,15 @@ double IcpPointToPoint::getResidual( double *T,const int32_t T_num,const Matrix 
 		double t0  = t.val[0][0]; double t1  = t.val[1][0];
 		for (int i=0; i<nact; i++) {
 			// kd tree query + result
-			std::vector<float>         query(m_dim);
+			std::vector<double>         query(m_dim);
 			kdtree::KDTreeResultVector result;
 
 			// get index of active point
 			int32_t idx = active[i];
 
 			// transform point according to R|t
-			query[0] = (float)(r00*T[idx*2+0] + r01*T[idx*2+1] + t0);
-			query[1] = (float)(r10*T[idx*2+0] + r11*T[idx*2+1] + t1);
+			query[0] = (r00*T[idx*2+0] + r01*T[idx*2+1] + t0);
+			query[1] = (r10*T[idx*2+0] + r11*T[idx*2+1] + t1);
 
 			// search nearest neighbor
 			m_kd_tree->n_nearest(query,1,result);
@@ -257,14 +257,14 @@ double IcpPointToPoint::getResidual( double *T,const int32_t T_num,const Matrix 
 		double t0  = t.val[0][0]; double t1  = t.val[1][0]; double t2  = t.val[2][0];
 		for (int i=0; i<nact; i++) {
 			// kd tree query + result
-			std::vector<float>         query(m_dim);
+			std::vector<double>         query(m_dim);
 			kdtree::KDTreeResultVector result;
 			// get index of active point
 			int32_t idx = active[i];
 			// transform point according to R|t
-			query[0] = (float)(r00*T[idx*3+0] + r01*T[idx*3+1] + r02*T[idx*3+2] + t0);
-			query[1] = (float)(r10*T[idx*3+0] + r11*T[idx*3+1] + r12*T[idx*3+2] + t1);
-			query[2] = (float)(r20*T[idx*3+0] + r21*T[idx*3+1] + r22*T[idx*3+2] + t2);
+			query[0] = (r00*T[idx*3+0] + r01*T[idx*3+1] + r02*T[idx*3+2] + t0);
+			query[1] = (r10*T[idx*3+0] + r11*T[idx*3+1] + r12*T[idx*3+2] + t1);
+			query[2] = (r20*T[idx*3+0] + r21*T[idx*3+1] + r22*T[idx*3+2] + t2);
 			// search nearest neighbor
 			m_kd_tree->n_nearest(query,1,result);
 			residual += result[0].dis;
